created 404 page in pages and errorContent in organism
