+++ src/pages/404.tsx
@@ -1,0 +1,16 @@
+import React, { FC } from 'react';
+import { Bar } from 'components/organisms';
+import { ErrorContent } from 'components/organisms/errorContent';
+
+const Custom404: FC = () => {
+  return (
+    <Bar>
+      <ErrorContent
+        title='404 Not Found'
+        subtitle='お探しのページが見つかりませんでした'
+      />
+    </Bar>
+  );
+};
+
+export default Custom404;
