+++ src/pages/profile/index.tsx
@@ -1,4 +1,4 @@
-import { ErrorPage } from 'components/templates/404Template';
+import { ErrorPage } from 'components/templates/errorTemplate';
 import { ProfileTemplate } from 'components/templates/profileTemplate';
 import { useAuth } from 'contexts/authContext';
 import React from 'react';
