+++ src/pages/random/index.tsx
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { ErrorPage } from 'components/templates/404Template';
+import { ErrorPage } from 'components/templates/errorTemplate';
 import { RandomTemplate } from 'components/templates/randomTemplate';
 import { AppContext } from 'contexts/appContext';
 import React, { useContext, useState } from 'react';
